+++ src/constants.rs
@@ -31,9 +31,9 @@ pub const FREQ_RANGE_PRESETS: [(u16, u16, &str); 5] = [
 ];
 
 // other
-pub const DEFAULT_CONFIG_DIR: &str = "audioleaf";
+pub const DEFAULT_CONFIG_DIR: &str = "nanoviz";
 pub const DEFAULT_CONFIG_FILE: &str = "config.toml";
 pub const DEFAULT_DEVICES_FILE: &str = "nl_devices.toml";
-pub const DEFAULT_BACKTRACE_FILE: &str = "audioleaf_backtrace.log";
+pub const DEFAULT_BACKTRACE_FILE: &str = "nanoviz_backtrace.log";
 pub const NL_API_PORT: u16 = 16021;
 pub const NL_UDP_PORT: u16 = 60222;